+++ src/pages/js/dark_light.js
@@ -1,10 +1,3 @@
-if (typeof browser === "undefined") {
-    browser = chrome
-    firefox = false
-}
-else{
-    firefox = true
-}
 let icon = document.querySelector('.sun-moon')
 function switchClass(element) {
     if (element.classList.contains("dark")) {
@@ -13,7 +6,7 @@ function switchClass(element) {
         document.querySelector("#d_l").innerHTML = "Dark"
         icon.classList.remove('fa-sun-bright')
         icon.classList.add('fa-moon')
-        browser.storage.local.set({mode: "light"})
+        chrome.storage.local.set({mode: "light"})
     }
     else if (element.classList.contains("light")) {
         element.classList.remove("light");
@@ -21,7 +14,7 @@ function switchClass(element) {
         document.querySelector("#d_l").innerHTML = "Light"
         icon.classList.remove('fa-moon')
         icon.classList.add('fa-sun-bright')
-        browser.storage.local.set({mode: "dark"})
+        chrome.storage.local.set({mode: "dark"})
     }
 
     // Recursively call the function for all children of the element
@@ -33,7 +26,7 @@ function toggleMode(){
     switchClass(document.body)
 }
 
-browser.storage.local.get({mode: "dark"}, function(result) {
+chrome.storage.local.get({mode: "dark"}, function(result) {
     if(result.mode === "light"){
         toggleMode()
     }