+++ src/pages/js/prompts.js
@@ -1,7 +1,3 @@
-if (typeof browser === "undefined") {
-    browser = chrome
-}
-
 let main = document.querySelector(".main");
 const modal = new bootstrap.Modal(document.getElementById('exploreModal'))
 
@@ -43,14 +39,14 @@ const keys_pressed = {
 
 let user_prompts = [];
 
-browser.storage.local.get({prompts: default_prompts}, function(result) {
+chrome.storage.local.get({prompts: default_prompts}, function(result) {
 	user_prompts = result.prompts;
 	load_prompts(user_prompts);
 });
 
 // sets up toggle for control save behavior
 let ctrlSave = false;
-browser.storage.local.get({settings: {ctrlSave: false}}, function(result){
+chrome.storage.local.get({settings: {ctrlSave: false}}, function(result){
 	ctrlSave = result.settings.ctrl_save;
 })
 
@@ -63,7 +59,7 @@ function searchString(string, searchTerm) {
 let dl;
 dark_light()
 async function dark_light() {
-	browser.storage.local.get({mode: "dark"},
+	chrome.storage.local.get({mode: "dark"},
 		function(result) {
 			dl = result?.mode;
 			if(!dl) dl = "dark"; // guard statement because it apparently still returns undefined "result" sometimes
@@ -223,7 +219,7 @@ Additional information:
 
 function delete_prompt(id)
 {
-	browser.storage.local.get({prompts: default_prompts}).then((result) => {
+	chrome.storage.local.get({prompts: default_prompts}).then((result) => {
 		let prompts = result.prompts;
 		let prompt = getObjectById(id, prompts);
 		if(!prompt)
@@ -233,10 +229,10 @@ function delete_prompt(id)
 		}
 		if (imported_prompts.includes(prompt.title)){
 			imported_prompts.splice(imported_prompts.indexOf(prompt.title), 1);
-			browser.storage.local.set({imported_prompts: imported_prompts});
+			chrome.storage.local.set({imported_prompts: imported_prompts});
 		}
 		removeElementInArray(prompts, prompt);
-		browser.storage.local.set({prompts: prompts});
+		chrome.storage.local.set({prompts: prompts});
 		user_prompts = prompts;
 		load_prompts(prompts);
 	});
@@ -245,7 +241,7 @@ function delete_prompt(id)
 function choose_category(id, row)
 {
 	let category = row.querySelector('.select').value;
-	browser.storage.local.get({prompts: default_prompts}).then((result) => {
+	chrome.storage.local.get({prompts: default_prompts}).then((result) => {
 		let prompts = result.prompts;
 		let prompt = getObjectById(id, prompts);
 		if(!prompt)
@@ -254,14 +250,14 @@ function choose_category(id, row)
 			return;
 		}
 		prompt.category = category
-		browser.storage.local.set({prompts: prompts});
+		chrome.storage.local.set({prompts: prompts});
 		load_prompts(prompts);
 	});
 }
 
 function use_prompt(id)
 {
-	browser.storage.local.get({prompts: default_prompts}).then((result) => {
+	chrome.storage.local.get({prompts: default_prompts}).then((result) => {
 		let prompts = result.prompts;
 		let prompt = getObjectById(id, prompts);
 		if(!prompt)
@@ -269,7 +265,7 @@ function use_prompt(id)
 			console.warn(`toggle_prompt_editable: cannot find prompt of id ${id}.`);
 			return;
 		}
-		browser.runtime.sendMessage({prompt: prompt.text, type: 'b_use_prompt'})
+		chrome.runtime.sendMessage({prompt: prompt.text, type: 'b_use_prompt'})
 	});
 }
 
@@ -285,7 +281,7 @@ function toggle_prompt_editable(id, element, just_title=false)
 		let textarea = document.createElement("textarea");
 		prompt_text.innerHTML = "";
 		prompt_text.appendChild(textarea)
-		browser.storage.local.get({prompts: default_prompts}).then((result) => {
+		chrome.storage.local.get({prompts: default_prompts}).then((result) => {
 			let prompts = result.prompts;
 			let prompt = getObjectById(id, prompts);
 			if(!prompt)